+++ deployment/updates/scripts/add_permissions/2025_07_22_update_webcast_producer_role.php
@@ -1,0 +1,4 @@
+<?php
+
+$script = realpath(dirname(__FILE__) . '/../../../../') . '/alpha/scripts/utils/permissions/addPermissionToRole.php';
+passthru("php $script 0 'WEbcast producer device role' 'ANALYTICS_BASE' realrun");