+++ plugins/reach/lib/model/HandlerFcatory.php
@@ -1,0 +1,29 @@
+<?php
+
+class HandlerFactory {
+
+	/**
+	 * @throws KalturaAPIException
+	 */
+	public static function getHandler($objectType): VendorTaskObjectHandler {
+		switch($objectType)
+		{
+			case EntryObjectType::ASSET:
+				return new AssetHandler();
+			case EntryObjectType::ENTRY:
+				return new EntryHandler();
+			default:
+				throw new KalturaAPIException(KalturaReachErrors::ENTRY_OBJECT_TYPE_NOT_SUPPORTED, $objectType);
+
+		}
+	}
+
+	/**
+	 * @throws KalturaAPIException
+	 */
+	public static function getHandlerByObject($object): VendorTaskObjectHandler
+	{
+		$objectType = $object instanceof AttachmentAsset ? EntryObjectType::ASSET : EntryObjectType::ENTRY;
+		return HandlerFactory::getHandler($objectType);
+	}
+}