+++ plugins/search/providers/elastic_search/lib/model/items/ESearchEntryItem.php
@@ -60,6 +60,7 @@ class ESearchEntryItem extends ESearchItem
 		'captions_content' => array('ESearchItemType::EXACT_MATCH'=> ESearchItemType::EXACT_MATCH, 'ESearchItemType::PARTIAL'=> ESearchItemType::PARTIAL, 'ESearchItemType::STARTS_WITH'=> ESearchItemType::STARTS_WITH, 'ESearchItemType::EXISTS'=> ESearchItemType::EXISTS, ESearchUnifiedItem::UNIFIED),
 		'last_played_at' => array('ESearchItemType::RANGE'=>ESearchItemType::RANGE, 'ESearchItemType::EXISTS'=> ESearchItemType::EXISTS),
 		'plays' => array('ESearchItemType::RANGE'=>ESearchItemType::RANGE),
+		'views' => array('ESearchItemType::RANGE'=>ESearchItemType::RANGE),
 		'rank' => array('ESearchItemType::RANGE'=>ESearchItemType::RANGE),
 		'votes' => array('ESearchItemType::RANGE'=>ESearchItemType::RANGE),
 		'display_in_search' => array('ESearchItemType::EXACT_MATCH' => ESearchItemType::EXACT_MATCH),